+++ ipns/record.go
@@ -12,6 +12,7 @@ import (
 	ipns_pb "github.com/ipfs/boxo/ipns/pb"
 	"github.com/ipfs/boxo/path"
 	"github.com/ipfs/boxo/util"
+	"github.com/ipfs/go-cid"
 	logging "github.com/ipfs/go-log/v2"
 	"github.com/ipld/go-ipld-prime"
 	"github.com/ipld/go-ipld-prime/codec/dagcbor"
@@ -30,6 +31,9 @@ type ValidityType int64
 // ValidityEOL means "this record is valid until {Validity}". This is currently
 // the only supported Validity type.
 const ValidityEOL ValidityType = 0
+
+// NoopValue is an identity CID that points at zero bytes.
+const NoopValue = "/ipfs/bafkqaaa"
 
 // Record represents an [IPNS Record].
 //
@@ -89,10 +93,29 @@ func (rec *Record) Value() (path.Path, error) {
 		return nil, err
 	}
 
-	p, err := path.NewPath(string(value))
+	if len(value) == 0 {
+		// To maximize interop across implementations, turn empty value into zero-length identity CID.
+		// This is a convenience placeholder used when the value in record is empty or does
+		// not matter because IPNS is used for custom CBOR in the Data field.
+		value = []byte(NoopValue)
+	}
+
+	// parse as a string with content path
+	if len(value) > 0 && value[0] == '/' {
+		p, err := path.NewPath(string(value))
+		if err != nil {
+			return nil, multierr.Combine(ErrInvalidPath, err)
+		}
+		// done, finish fast
+		return p, nil
+	}
+
+	// fallback: for legacy/optimization reasons, the value could be a valid CID in byte form
+	binaryCid, err := cid.Cast(value)
 	if err != nil {
 		return nil, multierr.Combine(ErrInvalidPath, err)
 	}
+	p := path.FromCid(binaryCid)
 
 	return p, nil
 }
@@ -232,6 +255,11 @@ func processOptions(opts ...Option) *options {
 // option [WithPublicKey]. In addition, records are, by default created with V1
 // compatibility.
 func NewRecord(sk ic.PrivKey, value path.Path, seq uint64, eol time.Time, ttl time.Duration, opts ...Option) (*Record, error) {
+	return newRecord(sk, []byte(value.String()), seq, eol, ttl, opts...)
+}
+
+// newRecord is a private version of NewRecord that allows arbitrary []byte values (used internally for testing)
+func newRecord(sk ic.PrivKey, value []byte, seq uint64, eol time.Time, ttl time.Duration, opts ...Option) (*Record, error) {
 	options := processOptions(opts...)
 
 	node, err := createNode(value, seq, eol, ttl)
@@ -260,7 +288,7 @@ func NewRecord(sk ic.PrivKey, value path.Path, seq uint64, eol time.Time, ttl ti
 	}
 
 	if options.v1Compatibility {
-		pb.Value = []byte(value.String())
+		pb.Value = value
 		typ := ipns_pb.IpnsRecord_EOL
 		pb.ValidityType = &typ
 		pb.Sequence = &seq
@@ -303,11 +331,11 @@ func NewRecord(sk ic.PrivKey, value path.Path, seq uint64, eol time.Time, ttl ti
 	}, nil
 }
 
-func createNode(value path.Path, seq uint64, eol time.Time, ttl time.Duration) (datamodel.Node, error) {
+func createNode(value []byte, seq uint64, eol time.Time, ttl time.Duration) (datamodel.Node, error) {
 	m := make(map[string]ipld.Node)
 	var keys []string
 
-	m[cborValueKey] = basicnode.NewBytes([]byte(value.String()))
+	m[cborValueKey] = basicnode.NewBytes(value)
 	keys = append(keys, cborValueKey)
 
 	m[cborValidityKey] = basicnode.NewBytes([]byte(util.FormatRFC3339(eol)))